[feat] 팔로우페이지 동적 라우팅 구현
- tab 상태 사용하지 않고 url 의 쿼리값으로 대체
- 동적라우팅 구현
- tab 클릭 이벤트에 shallow 라우팅 구현
